requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
scikit-learn
pandas
numpy
joblib
websockets
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
